FVA: HTTPS Rework - Added proper port mapping for http and https in docker-compose.yml. - Dev-cert now copied to container to make https possible. - Added command to generate dev cert. - Setup url in appsettings.Development.json - Added static endpoint to seed database. - We have E2E. - Setup httpclient to ignore ssl in development mode.
