+++ Client/OrganisationService.cs
@@ -6,6 +6,6 @@ public class OrganisationService(HttpClient client)
 {
     public async Task<OrganisationDto[]?> GetOrganisationsAsync()
     {
-        return await client.GetFromJsonAsync<OrganisationDto[]>("/api/organisations") ?? [];
+        return await client.GetFromJsonAsync<OrganisationDto[]>("api/Organisation") ?? [];
     }
 }
