+++ Client/appsettings.Development.json
@@ -1,8 +1,14 @@
 {
+  "https_port": 443,
   "Logging": {
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
-  }
+  },
+  "ConnectionStrings": {
+    "DefaultConnection": "User ID =postgres;Password=postgres;Host=127.0.0.1;Port=5432;Database=odin_db;Pooling=true;MinPoolSize=0;MaxPoolSize=20;Timeout=15;"
+  },
+  "server_host_address": "https://host.docker.internal:8081",
+  "DetailedErrors": true
 }
