+++ Client/appsettings.json
@@ -1,9 +1,13 @@
 {
+  "https_port": 443,
   "Logging": {
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "ConnectionStrings": {
+    "DefaultConnection": "User ID =postgres;Password=postgres;Host=127.0.0.1;Port=5432;Database=odin_db;Pooling=true;MinPoolSize=0;MaxPoolSize=20;Timeout=15;"
+  },
+  "server_host_address": "https://localhost:8081"
 }
