+++ Dockerfile-Client
@@ -4,16 +4,20 @@ EXPOSE 80
 EXPOSE 443
 
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+ARG ENV
+
 WORKDIR /src
 COPY ["/Client/Client.csproj", "Client/"]
 COPY ["/DTO/DTO.csproj", "DTO/"]
 RUN dotnet restore "Client/Client.csproj"
 COPY . .
-RUN dotnet build "DTO/DTO.csproj" -c Debug -o /app/build
-RUN dotnet build "Client/Client.csproj" -c Debug -o /app/build
+RUN dotnet build "DTO/DTO.csproj" -c ${ENV} -o /app/build
+RUN dotnet build "Client/Client.csproj" -c ${ENV} -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "Client/Client.csproj" -c Debug -o /app/publish
+ARG ENV
+
+RUN dotnet publish "Client/Client.csproj" -c ${ENV} -o /app/publish
 
 FROM base AS final
 WORKDIR /app
