+++ Dockerfile-Server
@@ -1,21 +1,25 @@
 ﻿FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
 USER $APP_UID
 WORKDIR /app
-EXPOSE 8080
+EXPOSE 80
+EXPOSE 443
 
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+ARG ENV
+
 WORKDIR /src
 COPY ["/Server/Server.csproj", "Server/"]
 COPY ["/DTO/DTO.csproj", "DTO/"]
 RUN dotnet restore "Server/Server.csproj"
 COPY . .
 WORKDIR "/src/"
-RUN dotnet build "DTO/DTO.csproj" -c Debug -o /app/build
-RUN dotnet build "Server/Server.csproj" -c Debug -o /app/build
+RUN dotnet build "DTO/DTO.csproj" -c ${ENV} -o /app/build
+RUN dotnet build "Server/Server.csproj" -c ${ENV} -o /app/build
 
 FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "Server/Server.csproj" -c Debug -o /app/publish /p:UseAppHost=false
+ARG ENV
+
+RUN dotnet publish "Server/Server.csproj" -c ${ENV} -o /app/publish /p:UseAppHost=false
 
 FROM base AS final
 WORKDIR /app
