+++ appsettings.json
@@ -1,4 +1,5 @@
 {
+  "https_port": "443",
   "Logging": {
     "LogLevel": {
       "Default": "Information",
