+++ docker-compose.yml
@@ -5,23 +5,37 @@
       - "server"
     container_name: fva_client
     ports:
-      - "8080:80"
+      - "5000:80"
+      - "5001:443"
     build:
       context: .
     environment:
-      - ASPNETCORE_URLS=http://+:80
+      - ASPNETCORE_ENVIRONMENT=Development
+      - ASPNETCORE_HTTPS_PORT=5001
+      - ASPNETCORE_URLS=https://+:443;http://+:80
+      - ASPNETCORE_Kestrel__Certificates__Default__Password=fva_password
+      - ASPNETCORE_Kestrel__Certificates__Default__Path=/https/aspnetapp.pfx
+    volumes:
+      - ${HOME}/.aspnet/https:/https/
   server:
     image: fva:server
+    ports:
+      - "8080:80"
+      - "8081:443"
     depends_on:
       - "app_db"
     container_name: fva_server
-    ports:
-      - "8081:80"
     build:
       context: .
     environment:
       - ConnectionStrings__DefaultConnection=User ID =postgres;Password=postgres;Server=app_db;Port=5432;Database=odin_db;Pooling=true;MinPoolSize=0;MaxPoolSize=20;Timeout=15;
-      - ASPNETCORE_URLS=http://+:80
+      - ASPNETCORE_ENVIRONMENT=Development
+      - ASPNETCORE_HTTPS_PORT=8081
+      - ASPNETCORE_URLS=https://+:443;http://+:80
+      - ASPNETCORE_Kestrel__Certificates__Default__Password=fva_password
+      - ASPNETCORE_Kestrel__Certificates__Default__Path=/https/aspnetapp.pfx
+    volumes:
+      - ${HOME}/.aspnet/https:/https/
   app_db:
     image: postgres:latest
     container_name: odin_db
