+++ server/Database/Models/ServiceLocation.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore;
+using Server.Database.Models;
 
 namespace FVA.Database.Models;
