+++ server/controllers/OrganisationController.cs
@@ -1,8 +1,4 @@
-using Bogus;
-using Bogus.DataSets;
-using DTO;
 using FVA.Database;
-using FVA.Database.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Server.Database.Models;
